+++ skimage/rank/_crank16.pyx
@@ -21,8 +21,8 @@ from _core16 cimport _core16
 # kernels uint16 take extra parameter for defining the bitdepth
 # -----------------------------------------------------------------
 
-cdef inline np.uint16_t kernel_autolevel(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i,imin,imax,delta
+cdef inline np.uint16_t kernel_autolevel(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i,imin,imax,delta
 
     if pop:
         for i in range(maxbin-1,-1,-1):
@@ -39,8 +39,8 @@ cdef inline np.uint16_t kernel_autolevel(int* histo, float pop, np.uint16_t g,in
     else:
         return <np.uint16_t>(imax-imin)
 
-cdef inline np.uint16_t kernel_bottomhat(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_bottomhat(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
 
     for i in range(maxbin):
         if histo[i]:
@@ -49,8 +49,8 @@ cdef inline np.uint16_t kernel_bottomhat(int* histo, float pop, np.uint16_t g,in
     return <np.uint16_t>(g-i)
 
 
-cdef inline np.uint16_t kernel_equalize(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_equalize(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
     cdef float sum = 0.
 
     if pop:
@@ -63,8 +63,8 @@ cdef inline np.uint16_t kernel_equalize(int* histo, float pop, np.uint16_t g,int
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_gradient(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i,imin,imax
+cdef inline np.uint16_t kernel_gradient(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i,imin,imax
 
     if pop:
         for i in range(maxbin-1,-1,-1):
@@ -79,8 +79,8 @@ cdef inline np.uint16_t kernel_gradient(int* histo, float pop, np.uint16_t g,int
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_maximum(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_maximum(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
 
     if pop:
         for i in range(maxbin-1,-1,-1):
@@ -89,8 +89,8 @@ cdef inline np.uint16_t kernel_maximum(int* histo, float pop, np.uint16_t g,int 
 
     return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_mean(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_mean(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
     cdef float mean = 0.
 
     if pop:
@@ -100,8 +100,8 @@ cdef inline np.uint16_t kernel_mean(int* histo, float pop, np.uint16_t g,int bit
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_meansubstraction(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_meansubstraction(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
     cdef float mean = 0.
 
     if pop:
@@ -111,8 +111,8 @@ cdef inline np.uint16_t kernel_meansubstraction(int* histo, float pop, np.uint16
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_median(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_median(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
     cdef float sum = pop/2.0
 
     if pop:
@@ -124,8 +124,8 @@ cdef inline np.uint16_t kernel_median(int* histo, float pop, np.uint16_t g,int b
 
     return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_minimum(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_minimum(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
 
     if pop:
         for i in range(maxbin):
@@ -134,8 +134,8 @@ cdef inline np.uint16_t kernel_minimum(int* histo, float pop, np.uint16_t g,int 
 
     return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_modal(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int hmax=0,imax=0
+cdef inline np.uint16_t kernel_modal(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t hmax=0,imax=0
 
     if pop:
         for i in range(maxbin):
@@ -146,8 +146,8 @@ cdef inline np.uint16_t kernel_modal(int* histo, float pop, np.uint16_t g,int bi
 
     return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_morph_contr_enh(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i,imin,imax
+cdef inline np.uint16_t kernel_morph_contr_enh(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i,imin,imax
 
     if pop:
         for i in range(maxbin-1,-1,-1):
@@ -165,11 +165,11 @@ cdef inline np.uint16_t kernel_morph_contr_enh(int* histo, float pop, np.uint16_
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_pop(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
+cdef inline np.uint16_t kernel_pop(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
     return <np.uint16_t>(pop)
 
-cdef inline np.uint16_t kernel_threshold(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_threshold(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
     cdef float mean = 0.
 
     if pop:
@@ -179,8 +179,8 @@ cdef inline np.uint16_t kernel_threshold(int* histo, float pop, np.uint16_t g,in
     else:
         return <np.uint16_t>(0)
 
-cdef inline np.uint16_t kernel_tophat(int* histo, float pop, np.uint16_t g,int bitdepth,int maxbin, int midbin):
-    cdef int i
+cdef inline np.uint16_t kernel_tophat(Py_ssize_t* histo, float pop, np.uint16_t g, Py_ssize_t bitdepth, Py_ssize_t maxbin, Py_ssize_t midbin):
+    cdef Py_ssize_t i
 
     for i in range(maxbin-1,-1,-1):
         if histo[i]:
@@ -195,7 +195,7 @@ def autolevel(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """bottom hat
     """
     return _core16(kernel_autolevel,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -204,7 +204,7 @@ def bottomhat(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """bottom hat
     """
     return _core16(kernel_bottomhat,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -213,7 +213,7 @@ def equalize(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local egalisation of the gray level
     """
     return _core16(kernel_equalize,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -222,7 +222,7 @@ def gradient(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local maximum - local minimum gray level
     """
     return _core16(kernel_gradient,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -231,7 +231,7 @@ def maximum(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local maximum gray level
     """
     return _core16(kernel_maximum,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -240,7 +240,7 @@ def mean(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """average gray level (clipped on uint8)
     """
     return _core16(kernel_mean,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -249,7 +249,7 @@ def meansubstraction(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """(g - average gray level)/2+midbin (clipped on uint8)
     """
     return _core16(kernel_meansubstraction,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -258,7 +258,7 @@ def median(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local median
     """
     return _core16(kernel_median,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -267,7 +267,7 @@ def minimum(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local minimum gray level
     """
     return _core16(kernel_minimum,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -276,7 +276,7 @@ def morph_contr_enh(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """morphological contrast enhancement
     """
     return _core16(kernel_morph_contr_enh,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -285,7 +285,7 @@ def modal(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """local mode
     """
     return _core16(kernel_modal,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -294,7 +294,7 @@ def pop(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """returns the number of actual pixels of the structuring element inside the mask
     """
     return _core16(kernel_pop,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -303,7 +303,7 @@ def threshold(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """returns maxbin-1 if gray level higher than local mean, 0 else
     """
     return _core16(kernel_threshold,image,selem,mask,out,shift_x,shift_y,bitdepth)
@@ -312,7 +312,7 @@ def tophat(np.ndarray[np.uint16_t, ndim=2] image,
             np.ndarray[np.uint8_t, ndim=2] selem,
             np.ndarray[np.uint8_t, ndim=2] mask=None,
             np.ndarray[np.uint16_t, ndim=2] out=None,
-            char shift_x=0, char shift_y=0, int bitdepth=8):
+            char shift_x=0, char shift_y=0, Py_ssize_t bitdepth=8):
     """top hat
     """
     return _core16(kernel_tophat,image,selem,mask,out,shift_x,shift_y,bitdepth)
